Refactor pod/booster shapes
